Fixed some VS warnings
Thanks to Fraggy again !


Former-commit-id: 4e4312862461949e554bd8ae24b352ed4e5ec734

+++ src/Nazara/Renderer/ShaderProgram.cpp
@@ -398,7 +398,7 @@ bool NzShaderProgram::LoadShaderFromFile(nzShaderType type, const NzString& file
 		return false;
 	}
 
-	unsigned int length = file.GetSize();
+	unsigned int length = static_cast<unsigned int>(file.GetSize());
 
 	NzString source;
 	source.Resize(length);
